// src/modules/dashboard/translate.enum.ts
export enum DashboardKeys {
    FINANCIAL_NUMBERS_AND_CHARTS_ERROR = "dashboard.FINANCIAL_NUMBERS_AND_CHARTS_ERROR",
    FINANCIAL_NUMBERS_ERROR = "dashboard.FINANCIAL_NUMBERS_ERROR",
    FINANCIAL_ACTUAL_VS_BUDGET_ERROR = "dashboard.FINANCIAL_ACTUAL_VS_BUDGET_ERROR",
    FINANCIAL_PL_BREAKDOWN_ERROR = "dashboard.FINANCIAL_PL_BREAKDOWN_ERROR",
    FINANCIAL_SALES_COMPONENT_ERROR = "dashboard.FINANCIAL_SALES_COMPONENT_ERROR",
    HR_NUMBERS_AND_CHARTS_ERROR = "dashboard.HR_NUMBERS_AND_CHARTS_ERROR",
    HR_NUMBERS_ERROR = "dashboard.HR_NUMBERS_ERROR",
    HR_STAFF_COSTS_YTD_ERROR = "dashboard.HR_STAFF_COSTS_YTD_ERROR",
    HR_STAFF_AVG_4_MONTHS_ERROR = "dashboard.HR_STAFF_AVG_4_MONTHS_ERROR",
    HR_ACTUAL_VS_BUDGET_ERROR = "dashboard.HR_ACTUAL_VS_BUDGET_ERROR",
}